refactor: reducing number of lines in trip packages
